Passing jubilee test

+++ src/core/transactions.js
@@ -87,10 +87,12 @@ function addDI ({ targets: { ledger, destination } }, { demurrage, income }) {
   ledger.income = income
   ledger.amount = ledger.income.subtract(ledger.demurrage)
   ledger.balance = ledger.balance.subtract(ledger.demurrage).add(ledger.income)
+  ledger.challenge = payload({ date: ledger.date, from: ledger, to: destination, amount: ledger.amount })
   destination.demurrage = ledger.demurrage.multiply(-1)
   destination.income = ledger.income.multiply(-1)
   destination.amount = destination.demurrage.add(destination.income)
   destination.balance = destination.balance.add(destination.demurrage).add(destination.income)
+  destination.challenge = payload({ date: ledger.date, from: destination, to: ledger, amount: destination.amount })
   return { ledger, destination }
 }
 /**
@@ -159,6 +161,7 @@ async function addSystemSignatures (table, { sources, targets }, keys) {
   if (!keys) {
     keys = KeyWrapper(await table.getItem({ ledger: 'system', entry: 'pk' }, 'System keys not found'))
   }
+  log(JSON.stringify(targets, null, 2))
   targets.destination.signature = await keys.privateKey.sign(targets.destination.challenge)
   const next = sha1(targets.destination.signature)
   targets.destination.next = next
@@ -200,7 +203,9 @@ function transition (table, { source, target }) {
   if (target.entry === 'pending' && isSigned(target)) {
     target.entry = '/current'
     table.putItem(target)
-    table.deleteItem({ ledger: target.ledger, entry: 'pending' })
+    if (target.ledger !== 'system') {
+      table.deleteItem({ ledger: target.ledger, entry: 'pending' })
+    }
     if (source.entry === '/current') {
       // bump to a permanent state
       source.entry = sortKey(source)
